requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dj-database-url==2.2.0
Django==5.1
djangorestframework==3.15.2
drf-yasg==1.21.7
flower==2.0.1
humanize==4.10.0
inflection==0.5.1
iniconfig==2.0.0
kombu==5.4.0
packaging==24.1
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pytest==8.3.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
setuptools==72.2.0
six==1.16.0
sqlparse==0.5.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
